Move TypedString
Move to base package to avoid being only class in package
