+++ modules/basics/src/main/java/com/opengamma/strata/basics/market/MarketDataFeed.java
@@ -9,7 +9,7 @@ import java.util.regex.Pattern;
 
 import org.joda.convert.FromString;
 
-import com.opengamma.strata.collect.type.TypedString;
+import com.opengamma.strata.collect.TypedString;
 
 /**
  * Identifies a feed of market data, for example Bloomberg or Reuters.
