+++ modules/calc/src/main/java/com/opengamma/strata/calc/config/pricing/FunctionGroupName.java
@@ -9,7 +9,7 @@ import java.util.regex.Pattern;
 
 import org.joda.convert.FromString;
 
-import com.opengamma.strata.collect.type.TypedString;
+import com.opengamma.strata.collect.TypedString;
 
 /**
  * The name of a function group.
