+++ modules/calc/src/main/java/com/opengamma/strata/calc/marketdata/config/MarketDataConfig.java
@@ -26,7 +26,7 @@ import org.joda.beans.impl.direct.DirectMetaPropertyMap;
 
 import com.google.common.collect.ImmutableMap;
 import com.opengamma.strata.collect.Messages;
-import com.opengamma.strata.collect.type.TypedString;
+import com.opengamma.strata.collect.TypedString;
 
 /**
  * Configuration required for building non-observable market data, for example curves or surfaces.
