--- modules/collect/src/main/java/com/opengamma/strata/collect/type/package-info.java
@@ -1,13 +1,0 @@
-/**
- * Copyright (C) 2014 - present by OpenGamma Inc. and the OpenGamma group of companies
- * 
- * Please see distribution for license.
- */
-
-/**
- * Data structures for types.
- * <p>
- * This package contains support for data types.
- */
-package com.opengamma.strata.collect.type;
-
