+++ modules/market/src/main/java/com/opengamma/strata/market/curve/CurveName.java
@@ -7,7 +7,7 @@ package com.opengamma.strata.market.curve;
 
 import org.joda.convert.FromString;
 
-import com.opengamma.strata.collect.type.TypedString;
+import com.opengamma.strata.collect.TypedString;
 
 /**
  * The name of a curve.
