+++ modules/market/src/main/java/com/opengamma/strata/market/surface/SurfaceName.java
@@ -7,7 +7,7 @@ package com.opengamma.strata.market.surface;
 
 import org.joda.convert.FromString;
 
-import com.opengamma.strata.collect.type.TypedString;
+import com.opengamma.strata.collect.TypedString;
 
 /**
  * The name of a surface.
