+++ modules/product/src/main/java/com/opengamma/strata/product/PositionAttributeType.java
@@ -7,7 +7,7 @@ package com.opengamma.strata.product;
 
 import org.joda.convert.FromString;
 
-import com.opengamma.strata.collect.type.TypedString;
+import com.opengamma.strata.collect.TypedString;
 
 /**
  * The type that provides meaning to a position attribute.
